backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
